examples/literacy_assessment/requirements.txt
# AWS Bedrock Knowledge Base access and chat models
boto3>=1.34.0
langchain-aws>=0.1.0

# LangChain core (using secure version)
langchain-core>=0.3.15

# Deep Agents framework
# Note: Install deepagents from parent project

# Data validation and serialization (using secure version)
pydantic>=2.4.0

# Environment variable management
python-dotenv>=1.0.0
